further cleaning up of mesh api (ints > floats)

+++ platform/include/PL/platform_mesh.h
@@ -94,7 +94,7 @@ PL_EXTERN_C
 PL_EXTERN PLMesh *plCreateMesh(PLMeshPrimitive primitive, PLMeshDrawMode mode, unsigned int num_tris, unsigned int num_verts);
 PL_EXTERN PLMesh *plCreateMeshInit( PLMeshPrimitive primitive, PLMeshDrawMode mode, unsigned int numTriangles, unsigned int numVerts,
 	const unsigned int* indicies, const PLVertex* vertices );
-PL_EXTERN PLMesh* plCreateMeshRectangle(int x, int y, unsigned int w, unsigned int h, PLColour colour);
+PL_EXTERN PLMesh* plCreateMeshRectangle( float x, float y, float w, float h, PLColour colour);
 PL_EXTERN void plDestroyMesh(PLMesh *mesh);
 
 PL_EXTERN void plDrawBevelledBorder( int x, int y, unsigned int w, unsigned int h );
